Final round of clean up
Use the Parser from dotnet/runtime and inject more methods into it using the partial class.
Mock out the DiagnosticDescriptors because we don't need localization.
