+++ src/Tools/ConfigurationSchemaGenerator/RuntimeSource/Configuration.Binder/Parser/KnownTypeSymbols.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
@@ -61,9 +61,9 @@ namespace Microsoft.Extensions.Configuration.Binder.SourceGeneration
         public INamedTypeSymbol? SerializationInfo { get; }
         public INamedTypeSymbol? IntPtr { get; }
         public INamedTypeSymbol? UIntPtr { get; }
-        public INamedTypeSymbol? MemberInfo { get; }
+        public INamedTypeSymbol? MemberInfo  { get; }
         public INamedTypeSymbol? ParameterInfo { get; }
-        public INamedTypeSymbol? Delegate { get; }
+        public INamedTypeSymbol? Delegate   { get; }
 
         public KnownTypeSymbols(CSharpCompilation compilation)
         {
